Add test comment feature
