+++ src/__mocks__/helpers.js
@@ -2,7 +2,7 @@ import {
   getItemsData, getComments, postComments, getLikes, likeRobot,
 } from '../api/api.js';
 
-const apiMock = [
+const likeApiMock = [
   {
     id: 1,
     name: 'Bender',
@@ -20,6 +20,41 @@ const apiMock = [
   },
 
 ];
+
+const commentsTest = (commentsArr) => {
+  document.body.innerHTML = '<ul id ="comments-ul"> </ul>';
+  const commentsUl = document.getElementById('comments-ul');
+
+  const commentList = document.createElement('li');
+  const commentArr = Array.from(document.querySelectorAll('#comments-ul li'));
+  commentsArr.forEach((element) => {
+    commentList.innerHTML = `${element.name}: ${element.comment}`;
+    commentsUl.appendChild(commentList);
+    commentArr.push(commentList);
+  });
+  return commentArr;
+};
+
+const commentApiMock = (id, name, userComment) => {
+  const commentsArr = [{
+    id: '1',
+    name: 'Kate',
+    userComment: 'new Comment 1',
+  }, {
+    id: '2',
+    name: 'Mike',
+    userComment: 'new Comment 2',
+  },
+  ];
+  const comment = {
+    id,
+    name,
+    userComment,
+  };
+  commentsArr.push(comment);
+  commentsTest(commentsArr);
+  return commentsArr;
+};
 
 const displaySection = document.getElementById('main-section__display-data');
 const robotsTotal = document.querySelector('#nav-title span');
@@ -161,4 +196,4 @@ function renderRobots() {
   });
 }
 
-export { apiMock, createRobots, renderRobots };
+export { likeApiMock, commentsTest, commentApiMock, createRobots, renderRobots };
