+++ src/test/homepage.test.js
@@ -2,18 +2,18 @@
  * @jest-environment jsdom
  */
 
-import { apiMock } from '../__mocks__/helpers.js';
+import { likeApiMock } from '../__mocks__/helpers.js';
 
-describe('count items from API', () => {
+describe('Retrieves likes from API', () => {
   test('count items from API', () => {
-    expect(apiMock.length).toBe(3);
+    expect(likeApiMock.length).toBe(3);
   });
 
   test('name from item 1', () => {
-    expect(apiMock[0].name).toBe('Bender');
+    expect(likeApiMock[0].name).toBe('Bender');
   });
 
   test('count likes from item 3', () => {
-    expect(apiMock[2].likes).toBe(20);
+    expect(likeApiMock[2].likes).toBe(20);
   });
 });
